refactor services

+++ src/services/receipt/ReceiptApi/Dtos/ReceiptDto.cs
@@ -5,6 +5,6 @@ public class ReceiptDto
     public int Id { get; set; }
     public string Message { get; set; } = null!;
     public int CustomerId { get; set; }
-    public long TicketId { get; set; }
+    public int TicketId { get; set; }
     public DateTime CreatedAt { get; set; }
 }
